+++ backend/src/main/java/lt/vtvpmc/threered/bookstore/DataInput.java
@@ -8,13 +8,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
-import lt.vtvpmc.threered.bookstore.model.Author;
-import lt.vtvpmc.threered.bookstore.model.Book;
-import lt.vtvpmc.threered.bookstore.model.Category;
-import lt.vtvpmc.threered.bookstore.model.Role;
-import lt.vtvpmc.threered.bookstore.model.User;
-import lt.vtvpmc.threered.bookstore.service.BookStoreService;
-import lt.vtvpmc.threered.bookstore.service.UserService;
+import lt.vtvpmc.threered.bookstore.author.Author;
+import lt.vtvpmc.threered.bookstore.book.Book;
+import lt.vtvpmc.threered.bookstore.book.BookStoreService;
+import lt.vtvpmc.threered.bookstore.category.Category;
+import lt.vtvpmc.threered.bookstore.role.Role;
+import lt.vtvpmc.threered.bookstore.user.User;
+import lt.vtvpmc.threered.bookstore.user.UserService;
 
 @Component
 public class DataInput implements CommandLineRunner {
