+++ backend/src/main/java/lt/vtvpmc/threered/bookstore/author/AuthorRepository.java
@@ -1,9 +1,7 @@
-package lt.vtvpmc.threered.bookstore.repository;
+package lt.vtvpmc.threered.bookstore.author;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-
-import lt.vtvpmc.threered.bookstore.model.Author;
 
 
 public interface AuthorRepository extends JpaRepository<Author, Long> {
