+++ backend/src/main/java/lt/vtvpmc/threered/bookstore/book/BookRepository.java
@@ -1,9 +1,8 @@
-package lt.vtvpmc.threered.bookstore.repository;
+package lt.vtvpmc.threered.bookstore.book;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import lt.vtvpmc.threered.bookstore.model.Book;
 
 
 public interface BookRepository extends JpaRepository<Book, Long> {
