+++ backend/src/main/java/lt/vtvpmc/threered/bookstore/category/CategoryRepository.java
@@ -1,10 +1,8 @@
-package lt.vtvpmc.threered.bookstore.repository;
+package lt.vtvpmc.threered.bookstore.category;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-
-import lt.vtvpmc.threered.bookstore.model.Category;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
