+++ backend/src/main/java/lt/vtvpmc/threered/bookstore/user/UserServiceImpl.java
@@ -1,13 +1,11 @@
-package lt.vtvpmc.threered.bookstore.service;
+package lt.vtvpmc.threered.bookstore.user;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import lt.vtvpmc.threered.bookstore.model.User;
-import lt.vtvpmc.threered.bookstore.repository.RoleRepository;
-import lt.vtvpmc.threered.bookstore.repository.UserRepository;
+import lt.vtvpmc.threered.bookstore.role.RoleRepository;
 
 import java.util.HashSet;
